feat: move logo to file
